+++ dvr_scan/config.py
@@ -313,20 +313,22 @@ _CONFIG_FILE_DIR: AnyStr = user_config_dir("DVR-Scan", False)
 
 USER_CONFIG_FILE_PATH: AnyStr = os.path.join(_CONFIG_FILE_DIR, _CONFIG_FILE_NAME)
 
+# TODO: Investigate if centralizing user help strings here would be useful.
+# It might make CLI help and documentation are easier to create, as well as
+# for generating a default config file template.
+
 # TODO: Replace these default values with those set in dvr_scan.context.
 CONFIG_MAP: ConfigDict = {
     # General Options
-    "region-editor": False,
     "quiet-mode": False,
-    "verbosity": "info",
-    "debug": False,
     # Input/Output
-    "input-mode": "opencv",
-    "output-dir": "",
-    "output-mode": "opencv",
     "ffmpeg-input-args": DEFAULT_FFMPEG_INPUT_ARGS,
     "ffmpeg-output-args": DEFAULT_FFMPEG_OUTPUT_ARGS,
+    "input-mode": "opencv",
     "opencv-codec": "XVID",
+    "output-dir": "",
+    "output-mode": "opencv",
+    "region-editor": False,
     # Motion Events
     "min-event-length": TimecodeValue("0.1s"),
     "time-before-event": TimecodeValue("1.5s"),
@@ -340,7 +342,7 @@ CONFIG_MAP: ConfigDict = {
     "kernel-size": KernelSizeValue(),
     "downscale-factor": 0,
     "learning-rate": float(-1),
-    # TODO(v1.7): Remove, replaced with region files.
+    # TODO(1.8): Remove, has been replaced with region files.
     "region-of-interest": RegionValueDeprecated(),
     "load-region": "",
     "frame-skip": 0,
@@ -361,12 +363,18 @@ CONFIG_MAP: ConfigDict = {
     "bounding-box-thickness": 0.0032,
     "bounding-box-min-size": 0.032,
     "thumbnails": None,
+    # Logging
+    "verbosity": "info",
+    "save-log": True,
+    "max-log-size": 20000,
+    "max-log-files": 4,
+    # Development
+    "debug": False,
 }
 """Mapping of valid configuration file parameters and their default values or placeholders.
 The types of these values are used when decoding the configuration file. Valid choices for
 certain string options are stored in `CHOICE_MAP`."""
 
-# TODO: This should be a validator. These sub- lists should also be constants somewhere.
 CHOICE_MAP: Dict[str, List[str]] = {
     "input-mode": ["opencv", "pyav", "moviepy"],
     "opencv-codec": ["XVID", "MP4V", "MP42", "H264"],